LayoutTests/fast/images/text-recognition/ios/coalesce-selection-rects-in-image-overlay-expected.txt

PASS selectionRectsAfterSelectingAll.length is 1
PASS successfullyParsed is true

TEST COMPLETE
